+++ src/tasks/minions/minigames/templeTrekkingActivity.ts
@@ -102,7 +102,7 @@ export default class extends Task {
 			user,
 			channelID,
 			str,
-			['trek', [quantity, difficulty], true],
+			['minigames', { temple_trek: { start: { difficulty, quantity } } }, true],
 			image!,
 			data,
 			itemsAdded
